rename DefaultDirective.php to DefaultFeatureGroup.php
