+++ src/FeatureGroups/DirectiveContract.php
@@ -1,6 +1,8 @@
-<?php declare(strict_types=1);
+<?php
 
-namespace Mazedlx\FeaturePolicy;
+declare(strict_types=1);
+
+namespace Mazedlx\FeaturePolicy\FeatureGroups;
 
 interface DirectiveContract
 {
