+++ src/FeatureGroups/FeatureGroupContract.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Mazedlx\FeaturePolicy\FeatureGroups;
+
+interface FeatureGroupContract
+{
+    public static function directive(string $directive): DirectiveContract;
+}
